+++ inGame.ino
@@ -1,7 +1,7 @@
 boolean isEndGame() 
 {
-   uint8_t dead = 0;
-  for(uint8_t i = 0 ; i< NB_MAX_ENNEMIES; ++i) {
+   int dead = 0;
+  for(int i = 0 ; i< NB_MAX_ENNEMIES; ++i) {
    if(ennemies[i].life>0) {
     if(ennemies[i].x< 2) {
       return true;
